#242 Query compliance
