+++ TAS_Engine/Query/GetDayTypeArray.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -27,32 +27,23 @@ using System.Text;
 using System.Threading.Tasks;
 using BHG = BH.oM.Geometry;
 using BHEE = BH.oM.Environment.Elements;
+
 using BH.Engine.Environment;
-using BH.oM.Geometry;
 
 namespace BH.Engine.TAS
 {
     public static partial class Query
     {
-        /***************************************************/
-        public static bool ElementIsOpening(this TBD.BuildingElementType tbdType)
+
+        private static TBD.dayType[] GetDayTypeArray(TBD.TBDDocument tbdDocument)
         {
-            switch (tbdType)
-            {
-                case TBD.BuildingElementType.ROOFLIGHT:
-                case TBD.BuildingElementType.DOORELEMENT:
-                case TBD.BuildingElementType.VEHICLEDOOR:
-                case TBD.BuildingElementType.GLAZING:
-                case TBD.BuildingElementType.CURTAINWALL:
-                case TBD.BuildingElementType.FRAMEELEMENT:
-                    return true;
-                case TBD.BuildingElementType.NOBETYPE:
-                case TBD.BuildingElementType.NULLELEMENT:
-                    return false;
-                default:
-                    return false;
-            }
+            TBD.dayType[] aDayTypeArray = new TBD.dayType[365];
+
+            TBD.Calendar aCalendar = tbdDocument.Building.GetCalendar();
+            for (int i = 0; i < 365; i++)
+                aDayTypeArray[i] = aCalendar.days(i + 1).dayType;
+
+            return aDayTypeArray;
         }
     }
 }
-
