+++ TAS_Engine/Query/GetInternalCondition.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -27,32 +27,30 @@ using System.Text;
 using System.Threading.Tasks;
 using BHG = BH.oM.Geometry;
 using BHEE = BH.oM.Environment.Elements;
+
 using BH.Engine.Environment;
-using BH.oM.Geometry;
 
 namespace BH.Engine.TAS
 {
     public static partial class Query
     {
-        /***************************************************/
-        public static bool ElementIsOpening(this TBD.BuildingElementType tbdType)
+        private static TBD.InternalCondition GetInternalCondition(TBD.zone zone, TBD.dayType dayType)
         {
-            switch (tbdType)
+            if (zone == null)
+                return null;
+
+
+            int aIndex = 0;
+            TBD.InternalCondition aInternalCondition = zone.GetIC(aIndex);
+            while (aInternalCondition != null)
             {
-                case TBD.BuildingElementType.ROOFLIGHT:
-                case TBD.BuildingElementType.DOORELEMENT:
-                case TBD.BuildingElementType.VEHICLEDOOR:
-                case TBD.BuildingElementType.GLAZING:
-                case TBD.BuildingElementType.CURTAINWALL:
-                case TBD.BuildingElementType.FRAMEELEMENT:
-                    return true;
-                case TBD.BuildingElementType.NOBETYPE:
-                case TBD.BuildingElementType.NULLELEMENT:
-                    return false;
-                default:
-                    return false;
+                if (HasDayType(aInternalCondition, dayType))
+                    return aInternalCondition;
+
+                aIndex++;
+                aInternalCondition = zone.GetIC(aIndex);
             }
+            return null; ;
         }
     }
 }
-
