+++ TAS_Engine/Query/GetZones.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -27,32 +27,31 @@ using System.Text;
 using System.Threading.Tasks;
 using BHG = BH.oM.Geometry;
 using BHEE = BH.oM.Environment.Elements;
+
 using BH.Engine.Environment;
-using BH.oM.Geometry;
 
 namespace BH.Engine.TAS
 {
     public static partial class Query
     {
-        /***************************************************/
-        public static bool ElementIsOpening(this TBD.BuildingElementType tbdType)
+        public static List<TBD.zone> GetZones(TBD.TBDDocument tbdDocument)
         {
-            switch (tbdType)
+            if (tbdDocument == null)
+                return null;
+
+            List<TBD.zone> aZoneList = new List<TBD.zone>();
+
+            int aIndex = 0;
+            TBD.zone aZone = tbdDocument.Building.GetZone(aIndex);
+            while (aZone != null)
             {
-                case TBD.BuildingElementType.ROOFLIGHT:
-                case TBD.BuildingElementType.DOORELEMENT:
-                case TBD.BuildingElementType.VEHICLEDOOR:
-                case TBD.BuildingElementType.GLAZING:
-                case TBD.BuildingElementType.CURTAINWALL:
-                case TBD.BuildingElementType.FRAMEELEMENT:
-                    return true;
-                case TBD.BuildingElementType.NOBETYPE:
-                case TBD.BuildingElementType.NULLELEMENT:
-                    return false;
-                default:
-                    return false;
+                aZoneList.Add(aZone);
+                aIndex++;
+                aZone = tbdDocument.Building.GetZone(aIndex);
             }
+
+            return aZoneList;
         }
+
     }
 }
-
